Add files via upload
This code gets the reflectance sensor to control the motor.
If the ref. sensor reads a dark color (2500) the motor stops.
If the ref. sensor reads a lighter color (not 2500) the motor keeps going.
This code also includes code for the distance sensor.
Uses switch cases to execute certain actions based off of the "state"